Moved StoreSelectedScope class from the Web to the Core project to allow referencing this type from other extension modules
